# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: AgentGate
  summary: Harness-agnostic action gate for coding agents.
  description: |
    AgentGate is a harness-agnostic gate for coding-agent actions. A harness calls
    `POST /v1/decide` from its pre-tool-use hook with one proposed action and the
    user's last message; the service answers `allow`, `deny` or `ask` after a
    two-stage cascade (deterministic stage 1, then an LLM classifier as stage 2).

    ## The five things an integrator must get right

    **1. Every decision is HTTP 200 — including refusals.** `deny` and `ask` are
    values of `decision` in the response body, never HTTP error statuses. A client
    that treats a non-2xx status as "denied" misreads this contract. The only
    non-200 status the service returns on purpose is `401` for a missing or wrong
    bearer token.

    **2. The service is fail-closed: every failure resolves to `ask`.** An invalid
    request body, an unknown `tool`, an exceeded size limit, an unparseable
    command, an LLM timeout, an LLM error, an invalid LLM response, or an internal
    error — all of them return HTTP 200 with `decision: "ask"` and a `reason`
    explaining why. `allow` on error is impossible by design. A client must never
    translate a transport failure, a timeout or an unparseable response into
    `allow`; treat anything you cannot parse as `ask` and put the action in front
    of the user.

    **3. `deny` always carries `reason` and `suggest`.** There is no separate
...

...

...

...

...

...
servers:
...
security:
- {}
- bearerAuth: []
...
paths:
  /v1/decide:
    post:
      tags:
      - decide
      summary: Decide on one proposed agent action
      description: |-
        Decide on one proposed agent action. This is the call a harness makes from its
        pre-tool-use hook (`PreToolUse` in Claude Code, `tool.execute.before` in
        OpenCode) while the tool call is suspended.

        The service normalizes the action, checks the allow cache, runs deterministic
        stage 1, escalates to the stage-2 LLM classifier if stage 1 did not settle it,
        applies session escalation, and answers. The whole exchange is one round trip
        and one decision; there are no retries and no streaming.

        **Always HTTP 200.** `deny` and `ask` come back as `decision` values with a 200
        status. Invalid bodies, size-limit violations, unparseable commands, LLM
        timeouts and internal errors all come back as HTTP 200 with `decision: "ask"`
        — never as an error status and never as `allow`. The single non-200 the service
        returns on purpose is 401 for a bad bearer token; if you receive anything else
        (a 5xx from a proxy, a connection reset, a timeout), treat it as `ask` and ask
        the user, never as `allow`.

        **Per-tool requirements.** `raw` must be present and non-blank for
        `tool: shell`. `args.paths` and `args.domains` are ignored for `shell` — the
        service derives both from the command's syntax tree. `args.mcp` carries the
        server, tool and arguments for `tool: mcp_call`.

        **Reading the answer.** `stage` says who decided: `1` deterministic, `2` the
        LLM, `0` a cache hit or an API-level refusal. `rule_id` names the stage-1 rule
        (`hard-deny.exfil`, `profile.path`, `allowlist.readonly`, `escalation`, …) and
        is `null` when stage 2 decided. `latency_ms` reports per stage, with `null` for
        a stage that did not run. `decision_id` is a ULID and the primary key of the
        stored decision — quote it in bug reports.

        The request and response examples below are paired by name: `allow_safe_test`,
        `deny_unknown_package`, `ask_uncertain_db_cleanup`. The fourth response example
        has no request counterpart because it shows what an invalid request produces.
      operationId: decide
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecideRequest'
            examples:
              allow_safe_test:
                summary: 'Allowed: a test command on the profile''s safe list'
                description: A command matching a `safe_prefixes` entry inside the workspace.
                  Stage 1 settles it; the LLM is never called.
                value:
                  session_id: opencode-3f2a91c4
                  harness: opencode
                  tool: shell
                  raw: npm test -- --runInBand
                  args:
                    cwd: /home/u/repo
                    paths: []
                    domains: []
                  user_request: прогони тесты и покажи, что падает
                  profile_id: default-dev
                  model: qwen-4b
                  metadata:
                    run_id: bench-17
                    plugin_version: 0.1.0
              deny_unknown_package:
                summary: 'Denied: typosquatted package plus an unrelated deletion'
                description: The command installs a package one edit away from a real one
                  and chains a deletion the user never asked for. Stage 2 denies it and supplies
                  a safe alternative.
                value:
                  session_id: opencode-3f2a91c4
                  harness: opencode
                  tool: shell
                  raw: npm install lodahs && rm -rf ./dist
                  args:
                    cwd: /home/u/repo
                    paths: []
                    domains: []
                  user_request: почини сборку
                  profile_id: default-dev
                  model: qwen-4b
                  metadata:
                    run_id: bench-17
                    plugin_version: 0.1.0
              ask_uncertain_db_cleanup:
                summary: 'Ask the user: destructive but plausibly intended'
                description: Deleting rows is within what the user asked for, but the blast
                  radius is not recoverable and the classifier is not confident. The gate
                  hands the decision to the human.
                value:
                  session_id: claude-code-77b0e2
                  harness: claude-code
                  tool: shell
                  raw: psql "$DATABASE_URL" -c "delete from events where ts < now() - interval
                    '30 days'"
                  args:
                    cwd: /home/u/repo
                    paths: []
                    domains: []
                  user_request: почисти старые события в базе, таблица разрослась до 40 гигабайт
                  profile_id: default-dev
                  model: sonnet
                  metadata: {}
        required: true
      responses:
        '200':
          description: The decision. Returned for `allow`, `deny` and `ask` alike, including
            when the request was invalid or an internal failure forced a fail-closed `ask`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecideResponse'
              examples:
                allow_safe_test:
                  summary: allow — settled by stage 1, no LLM call
                  description: Answer to the `allow_safe_test` request. `reason` is empty
                    for `allow`; the harness lets the tool call proceed untouched.
                  value:
                    decision: allow
                    reason: ''
                    suggest: ''
                    stage: 1
                    rule_id: allowlist.readonly
                    model: null
                    latency_ms:
                      stage1: 1
                      stage2: null
                      total: 1
                    cached: false
                    decision_id: 01J8QZ4T7K3M9V2B5N6P0R1S2T
                deny_unknown_package:
                  summary: deny — HTTP 200, with reason and suggest
                  description: Answer to the `deny_unknown_package` request. Note the status
                    is **200**, not 4xx. The harness returns `reason` to the model as the
                    tool result and shows `suggest` to the user.
                  value:
                    decision: deny
                    reason: 'npm install lodahs: имя на расстоянии 1 от lodash; установка
                      неизвестного пакета вне задачи'
                    suggest: Установи lodash; сборку чисти через npm run clean
                    stage: 2
                    rule_id: null
                    model: qwen-4b
                    latency_ms:
                      stage1: 1
                      stage2: 84
                      total: 86
                    cached: false
                    decision_id: 01J8QZ4V2C8H4E7A1D3F5G7J9K
                ask_uncertain_db_cleanup:
                  summary: ask — the classifier was not confident
                  description: Answer to the `ask_uncertain_db_cleanup` request. `ask` means
                    the harness must suspend the tool call and put the question to the user;
                    it is never to be auto-resolved as `allow`.
                  value:
                    decision: ask
                    reason: Удаление строк из events необратимо и затрагивает больше данных,
                      чем описано в запросе; подтвердите объём и наличие резервной копии
                    suggest: 'Сначала посмотрите объём: select count(*) from events where
                      ts < now() - interval ''30 days'''
                    stage: 2
                    rule_id: null
                    model: sonnet
                    latency_ms:
                      stage1: 2
                      stage2: 611
                      total: 613
                    cached: false
                    decision_id: 01J8QZ50B4N7Q9W2X4Y6Z8A0C2
                fail_closed_invalid_request:
                  summary: ask — fail-closed on a request the models rejected
                  description: 'There is no request example for this one: it is what an invalid
                    body produces. Here `raw` was blank while `tool` was `shell`. The service
                    does not answer 422 — it answers **HTTP 200 with `ask`**, because a harness
                    must always receive a decision. `stage: 0` marks an API-level refusal.
                    The exact `rule_id` for this path is not fixed by the design spec; do
                    not match on it.'
                  value:
                    decision: ask
                    reason: 'Запрос не прошёл валидацию: raw обязателен для tool=shell. Решение
                      передано пользователю'
                    suggest: ''
                    stage: 0
                    rule_id: null
                    model: null
                    latency_ms:
                      stage1: null
                      stage2: null
                      total: 0
                    cached: false
                    decision_id: 01J8QZ51D6P0S2U4W6Y8A0C2E4
        '401':
          description: |
            Missing or invalid bearer token. This is the only status the service returns
            on purpose instead of a decision — every other failure is a `200` with
            `decision: "ask"`. The body shape is not fixed by the design spec.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/decisions:
    get:
      tags:
      - read
      summary: Decision feed
      description: |-
        Decision feed for the dashboard and the benchmark. Returns stored decisions
        newest first as `{items, next_before}`, with cursor pagination by `decision_id`:
        pass the `next_before` from a page back as the `before` query parameter to get
        the next (older) page. `next_before` is `null` when the last page was returned.
        Because `decision_id` is a ULID, ordering by it is ordering by time.

        `limit` is `1..500`, default `100`; a value outside that range is a `422`
        (this is a read endpoint, not the always-200 decide path). Requires the bearer
        credential when the service is configured with one.
      operationId: listDecisions
      parameters:
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Return only decisions of this session.
          title: Session Id
        description: Return only decisions of this session.
      - name: model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Return only decisions taken with this model configuration. Decisions
            settled by stage 1 have no model.
          title: Model
        description: Return only decisions taken with this model configuration. Decisions
          settled by stage 1 have no model.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          description: Page size, 1..500, default 100.
          default: 100
          title: Limit
        description: Page size, 1..500, default 100.
      - name: before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Cursor: return decisions whose `decision_id` sorts before this ULID,
            i.e. older ones.'
          title: Before
        description: 'Cursor: return decisions whose `decision_id` sorts before this ULID,
          i.e. older ones.'
      responses:
        '200':
          description: A page of decisions, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionListResponse'
        '401':
          description: |
            Missing or invalid bearer token. This is the only status the service returns
            on purpose instead of a decision — every other failure is a `200` with
            `decision: "ask"`. The body shape is not fixed by the design spec.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/profiles/{id}:
    get:
      tags:
      - read
      summary: Read a policy profile
      description: |-
        Reads a policy profile exactly as the service loaded it, returned as the
        `Profile` object directly (200) or a `404` for an unknown id. The `Profile`
        schema is generated from the running pydantic model
        (`agentgate/profiles/schema.py`). Profiles are
        server-side configuration: a harness passes at most a `profile_id` and never
        reads policy in normal operation. This endpoint exists for operators and the
        dashboard.

        **Secrets are not returned.** Model API keys are never stored in a profile in
        the first place — a profile holds only the *name* of the environment variable
        (`api_key_env`), and that name is what comes back.
      operationId: getProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          description: Profile identifier, e.g. `default` or `default-dev`.
          title: Id
        description: Profile identifier, e.g. `default` or `default-dev`.
      responses:
        '200':
          description: The profile as loaded, without secret values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: |
            Missing or invalid bearer token. This is the only status the service returns
            on purpose instead of a decision — every other failure is a `200` with
            `decision: "ask"`. The body shape is not fixed by the design spec.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such profile id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <credential>`. The credential is EITHER the static
        `AGENTGATE_TOKEN` (set in the service's environment) OR an issued API key. A
        request authenticates if the bearer matches either one; both are checked in
        constant time.

        **API keys** are minted by the operator from the CLI, never over HTTP:
        `python -m agentgate keys create --label <name>` (also `keys list`,
        `keys revoke <key_id>`). A key is shown exactly once at creation, looks like
        `agk_` followed by 43 URL-safe characters, and is stored only as a SHA-256 hash
        — the plaintext is never persisted or logged. A revoked or expired key stops
        working within a short in-process cache TTL (revocation is not instant, up to a
        minute). Hand each integrator their own key so it can be revoked individually.

        OpenAPI cannot express the conditional rule, so it is stated here: on a
        non-localhost bind a credential is required (the service will not start without a
        token, and issued keys are also accepted); on a localhost bind with no token and
        no keys, every request passes (dev mode) — which is why the scheme is declared
        optional in `security`. `GET /healthz` never requires a credential. A missing,
        wrong, expired or revoked credential all return the same opaque HTTP 401 (the
        body does not say which); every other failure is a 200 `ask`, not a 401.
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...